Add autotiled roads and building icons
Currently the icons are implemented in a hacky way since I incremented
all the building IDs by 10 to leave space for tiles, meaning I had to
create 9 dummy tiles and buildings. This might be fixable with a tile
atlas, or perhaps a dict that maps palette tile IDs to buildings.