Start migrating to Godot 4 (WIP)
This migration is in very early stages. The main menu is mostly working
and the main scene works at a very basic level without crashing. Bugs
abound!

Migrating this project will be difficult. When we made Minieval, it was
our first time using Godot, so the project is full of anti-patterns. Our
custom TileMap code in particular will be difficult to port over,
especially with the changes made to TileMap in Godot 4. There are lots
of cool new features, but TileMaps seem even harder to work with in code
now.

The main obstacle right now is road terrains. After that, there are a
large number of (hopefully isolated) issues to fix. The in-game menus
need to be rewritten, preferably in an event-driven style. Many other
scripts could stand to be similarly refactored. This would support
better scene decomposition too. RichTextLabels need to be fixed. I think
Camera zoom is partially broken. TileMap could use major refactoring.

There is a good chance that I will abandon this migration, but I at
least want to save the progress I've made so far.